First Commit
Setup necessary classes for trading crypto. Capable of buying, and selling, while maintaining persistence across sessions.

Successful purchases are stored in "inventory" allowing for specific orders to be sold.
Orders from inventory can be sold by providing the order id of the purchased order. If no order id is provided, the oldest purchase is used. When a sell order is closed, the original purchase is reinserted to inventory if the sell order was not 100% filled.

The coinbase websocket is used for live ticker data, while the rest client is used to track and place orders. Orders are tracked in a loop to ensure that no order is lost in case of session closure. When a new session starts, any previous orders are placed back into the tracking loop to catch all changes.