Autoformer For Decisions
Extensively trained a Autoformer NN using neural forecast. The model's outputs are trivial, seemingly random floats.
Upon testing the model against both live, and historical data, there seemed to be a pattern in its predictions. This indicated to me that the model does in fact have an understanding of the data.
To use this predictions as intended, I needed to determine how I can convert each raw value into one of three integers: -1, 0, or 1 (sell, hold, buy).
Within the NF_DecisionAutoformer, I wrote a program to extract frequently occuring predictions, and map these predictions to their corresponding integer.
The results of this are stored inside of base_predict as the variable "DECISION_TRANSLATIONS".
The code to get these translations from raw predictions can be found in base_predict.translate_decision